Rag/Project/requirements.txt
# Core web framework
fastapi==0.115.5
uvicorn==0.30.6

# Template rendering
jinja2==3.1.4
python-multipart==0.0.9

# Environment variable handling
python-dotenv==1.0.1

# LlamaIndex (for RAG + embeddings)
llama-index==0.11.16
llama-index-embeddings-openai==0.2.5
llama-index-llms-openai==0.2.11
llama-index-vector-stores-chroma==0.2.2

# Chroma for vector storage
chromadb==0.5.5

# Image and document handling (optional)
pillow==10.4.0

# Optional CLI tool
typer==0.12.5
